pyproject.toml

[project]
name = "verve-backend"
version = "1.5.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
  "alembic>=1.17.2",
  "bcrypt==4.0.1",
  "boto3>=1.40.14",
  "boto3-stubs[s3]>=1.40.21",
  "celery>=5.5.3",
  "fastapi[standard]>=0.115.12",
  "geo-track-analyzer @ git+https://github.com/kschweiger/track_analyzer@v2",
  "geoalchemy2>=0.17.1",
  "gevent>=25.9.1",
  "passlib>=1.7.4",
  "psycopg>=3.2.12",
  "pydantic>=2.11.4",
  "pydantic-settings>=2.9.1",
  "pyjwt>=2.10.1",
  "pyproj>=3.7.1",
  "redis>=7.0.1",
  "shapely>=2.1.2",
  "sqlalchemy>=2.0.41",
  "sqlmodel>=0.0.24",
  "structlog>=25.5.0",
]

[project.scripts]
setup-db = "verve_backend.cli.setup_db:main"
create-admin = "verve_backend.cli.create_admin_user:main"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
  "E",
  "F",
  "I",
  "PD",
  "PT",
  "NPY",
  "N",
  "W",
  "RET",
  "ANN",
  "RUF",
  "SIM",
  "PERF",
]
unfixable = ["F841", "RUF015"]
ignore = [
  "ANN202",
  "ANN205",
  "ANN001",
  "ANN002",
  "ANN003",
  "ANN401",
  "RET502",
  "RET503",
  "RET504",
  "RET505",
  "RUF005",
  "RUF009",
  "E731",
]

[dependency-groups]
dev = [
  "debugpy>=1.8.17",
  "matplotlib>=3.10.3",
  "pandas>=2.2.3",
  "plotly>=5.24.1",
  "rich>=14.2.0",
  "seaborn>=0.13.2",
  { include-group = "test" },
]
test = [
  "pillow>=12.0.0",
  "pytest>=8.4.2",
  "pytest-cov>=7.0.0",
  "pytest-mock>=3.15.1",
  "pytest-sugar>=1.1.1",
]
[tool.uv]
package = true

[build-system]
requires = ["uv_build>=0.8.13,<0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "verve_backend"
module-root = ""

[tool.git-changelog]
output = "CHANGELOG.md"
convention = "conventional"
template = "angular"
